Improve update_file and add update_audio_file

+++ backend/webserver/db.py
@@ -33,6 +33,9 @@ client = session.client('s3',
 BUCKET = 'ultracast-files'
 FILE_ACCESS = 'public-read'
 
+AUDIO_FORMAT_TYPE = "audio/mpeg"
+AUDIO_EXTENSION = ".mp3"
+
 def get_bucket_url():
     return re.sub(r"^https://", f"https://{BUCKET}.", STATIC_FILE_BASE_URL)
 
@@ -56,9 +59,14 @@ def file_exists(key):
     except:
         return False
 
-def add_file(data, key=None, override=False, content_type=None, ext=""):
+def get_key(data, key=None, ext=""):
     if key is None:
-        key = get_key_from_binary_data(data, ext)
+        return get_key_from_binary_data(data, ext)
+    else:
+        return key
+
+def add_file(data, key=None, override=False, content_type=None, ext=""):
+    key = get_key(data, key, ext)
     
     if not override and file_exists(key):
         return get_file_url(key)
@@ -73,12 +81,16 @@ def add_file(data, key=None, override=False, content_type=None, ext=""):
     return get_file_url(key)
 
 def add_audio_file(data, key=None, override=False):
-    return add_file(data, key, override, "audio/mpeg", ".mp3")
+    return add_file(data, key, override, AUDIO_FORMAT_TYPE, AUDIO_EXTENSION)
 
 def remove_file(url):
     resp = client.delete_object(Bucket=BUCKET, Key=get_key_from_url(url))
     check_status(resp, [200, 204], 'Remove File')
 
-def update_file(old_url, data, new_key=None, content_type=None):
-    remove_file(old_url)
-    return add_file(data, new_key, content_type=content_type)
+def update_file(old_url, data, new_key=None, content_type=None, ext=""):
+    if file_exists(get_key_from_url(old_url)):
+        remove_file(old_url)
+    return add_file(data, new_key, content_type=content_type, ext=ext)
+
+def update_audio_file(old_url, data, new_key=None):
+    return update_file(old_url, data, new_key, AUDIO_FORMAT_TYPE, AUDIO_EXTENSION)
